fix(nip66): address relay probe review feedback
